+++ App/Sources/Screens/NumbersRunView+UITest.swift
@@ -82,7 +82,8 @@ extension NumbersRunState {
                hintUsed: hintUsed,
                feedback: feedback ?? self.feedback,
                otherWord: self.otherWord,
-               finished: finished)
+               finished: finished,
+               score: score)
     }
 }
 #endif