+++ android/src/main/kotlin/net/spross/app/TrainerNaming.kt
@@ -7,6 +7,7 @@ import net.spross.kern.trainer.DrillModifier
 import net.spross.kern.trainer.LetterStage
 import net.spross.kern.trainer.Numbers
 import net.spross.kern.trainer.NumbersExercise
+import net.spross.kern.trainer.TimedRun
 import net.spross.kern.trainer.numbersExerciseEmoji
 
 /**
@@ -35,12 +36,14 @@ fun Chrome.name(modifier: DrillModifier): String = when (modifier) {
     DrillModifier.Reverse -> trainerModifierReverse
     DrillModifier.Fast -> trainerModifierFast
     DrillModifier.Mix -> trainerModifierMix
+    DrillModifier.Timed -> trainerModifierTimed
 }
 
 fun Chrome.hint(modifier: DrillModifier): String = when (modifier) {
     DrillModifier.Reverse -> trainerModifierReverseHint
     DrillModifier.Fast -> trainerModifierFastHint.format(Numbers.winsToAdvance(fast = false))
     DrillModifier.Mix -> trainerModifierMixHint
+    DrillModifier.Timed -> trainerModifierTimedHint.format(TimedRun.SECONDS)
 }
 
 fun Chrome.name(stage: LetterStage): String = when (stage) {